Refactor and remove redundant code for 2024
