+++ src/main/java/com/disastrousdata/AutoState.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package com.disastrousdata;
 
 public class AutoState {
     public TankDrive drive;
