+++ src/main/java/com/disastrousdata/Command.java
@@ -1,0 +1,5 @@
+package com.disastrousdata;
+
+public interface Command {
+    HardwareStates Execute(TankDrive drive, HardwareStates states);
+}
