+++ src/main/java/com/disastrousdata/HardwareStates.java
@@ -1,0 +1,6 @@
+package com.disastrousdata;
+
+public class HardwareStates {
+    public double LeftDriveMotors = 0;
+    public double RightDriveMotors = 0;
+}
