+++ src/main/java/com/disastrousdata/Main.java
@@ -1,15 +1,12 @@
-// Copyright (c) Rowville Roobots Team
+// Copyright (c) Disastrous Data
 // Contributors (In order of contribution): CoPokBl, Camo
 
-package frc.robot;
+package com.disastrousdata;
 
 import edu.wpi.first.wpilibj.RobotBase;
 
-// This class should not be modified
 public final class Main {
   private Main() {}
-
-  // This should only be changed if the Robot class is renamed or moved
   public static void main(String... args) {
     RobotBase.startRobot(Robot::new);
   }
