--- src/main/java/frc/robot/ButtonEvent.java
@@ -1,5 +1,0 @@
-package frc.robot;
-
-public class ButtonEvent {
-    
-}
