--- src/main/java/frc/robot/Command.java
@@ -1,5 +1,0 @@
-package frc.robot;
-
-public interface Command {
-    public HardwareStates Execute(TankDrive drive, HardwareStates states);
-}
