--- src/main/java/frc/robot/HardwareStates.java
@@ -1,14 +1,0 @@
-package frc.robot;
-
-public class HardwareStates {
-    public double LeftDriveMotors = 0;
-    public double RightDriveMotors = 0;
-
-    public double Arm = 0;
-
-    public double ArmInOutMotors = 0;
-
-    public double Intake = 0;
-
-    // public boolean ClawPiston = false;
-}
